Changed order, showed loss and made an extra slide;
I think it would be a good idea to leave the final results to end
as it follows convention of "it's all about the journey, not the destination".
Also IMO it feels more natural to first talk what we did and how, and then
show that the final model isn't performing that nicely, even though
the things we tried felt reasonable and there were some logic behind
creating the model. Also in Moodle, they aren't directly saying
the mandatory order. But I'm still quite fine if we would
throw the results as first part :P